+++ src/types/config.ts
@@ -79,6 +79,7 @@ export default interface Config {
     scrobbleTimeline: {
       fetchScrobbles: Script;
       mergeScrobbles: Script;
+      groupArtistsByGenres: Script;
     };
   };