+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+#### Scrobble timeline
+* A dataset that maps artists to their main genres.
+* A script that combines this dataset into a single file.
 
 ## [1.2.0] - 2019-11-09
 ### Added